# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS API",
    "version": "1.0.0",
    "description": "Minimal WMS API with Swagger UI"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login and get JWT token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin.demo@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT issued"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/product-lots": {
      "get": {
        "summary": "List product lots",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "client_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/warehouses": {
      "get": {
        "summary": "List warehouses",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create warehouse",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "name",
                  "country"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "timezone": {
                    "type": "string",
                    "example": "Asia/Bangkok"
                  },
                  "status": {
                    "type": "string",
                    "example": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Duplicate code"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "List products",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_id",
                  "barcode_raw",
                  "barcode_full",
                  "name_kr"
                ],
                "properties": {
                  "client_id": {
                    "type": "integer"
                  },
                  "sku_code": {
                    "type": "string"
                  },
                  "barcode_raw": {
                    "type": "string"
                  },
                  "barcode_full": {
                    "type": "string"
                  },
                  "name_kr": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string"
                  },
                  "volume_ml": {
                    "type": "integer"
                  },
                  "unit": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "example": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          },
          "409": {
            "description": "Duplicate barcode"
          }
        }
      }
    },
...
    "/clients": {
      "get": {
        "summary": "List clients",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create client",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_code",
                  "name_kr"
                ],
                "properties": {
                  "client_code": {
                    "type": "string"
                  },
                  "name_kr": {
                    "type": "string"
                  },
                  "name_en": {
                    "type": "string"
                  },
                  "contact_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "example": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Duplicate client_code"
          }
        }
      }
    },
...
    "/inbound-orders": {
      "get": {
        "summary": "List inbound orders",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create inbound order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "inbound_no",
                  "client_id",
                  "warehouse_id",
                  "inbound_date",
                  "created_by"
                ],
                "properties": {
                  "inbound_no": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "integer"
                  },
                  "warehouse_id": {
                    "type": "integer"
                  },
                  "inbound_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "example": "draft"
                  },
                  "memo": {
                    "type": "string"
                  },
                  "created_by": {
                    "type": "integer"
                  },
                  "received_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          },
          "409": {
            "description": "Duplicate inbound_no"
          }
        }
      }
    },
...
    "/inbound-items": {
      "get": {
        "summary": "List inbound items",
        "parameters": [
          {
            "name": "inbound_order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create inbound item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "inbound_order_id",
                  "product_id",
                  "lot_id",
                  "qty"
                ],
                "properties": {
                  "inbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "invoice_price": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "example": "THB"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          }
        }
      }
    },
...
    "/outbound-orders": {
      "get": {
        "summary": "List outbound orders",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create outbound order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "outbound_no",
                  "client_id",
                  "warehouse_id",
                  "order_date",
                  "created_by"
                ],
                "properties": {
                  "outbound_no": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "integer"
                  },
                  "warehouse_id": {
                    "type": "integer"
                  },
                  "order_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "sales_channel": {
                    "type": "string"
                  },
                  "order_no": {
                    "type": "string"
                  },
                  "tracking_no": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "example": "draft"
                  },
                  "packed_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "shipped_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "created_by": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          },
          "409": {
            "description": "Duplicate outbound_no"
          }
        }
      }
    },
...
    "/outbound-items": {
      "get": {
        "summary": "List outbound items",
        "parameters": [
          {
            "name": "outbound_order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create outbound item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "outbound_order_id",
                  "product_id",
                  "lot_id",
                  "qty"
                ],
                "properties": {
                  "outbound_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "box_type": {
                    "type": "string"
                  },
                  "box_count": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          }
        }
      }
    },
...
    "/return-orders": {
      "get": {
        "summary": "List return orders",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create return order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "return_no",
                  "client_id",
                  "warehouse_id",
                  "return_date",
                  "created_by"
                ],
                "properties": {
                  "return_no": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "integer"
                  },
                  "warehouse_id": {
                    "type": "integer"
                  },
                  "related_outbound_order_id": {
                    "type": "integer"
                  },
                  "return_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "example": "draft"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "created_by": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          },
          "409": {
            "description": "Duplicate return_no"
          }
        }
      }
    },
...
    "/return-items": {
      "get": {
        "summary": "List return items",
        "parameters": [
          {
            "name": "return_order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create return item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "return_order_id",
                  "product_id",
                  "lot_id",
                  "qty_received"
                ],
                "properties": {
                  "return_order_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "lot_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "integer"
                  },
                  "qty_received": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "qty_restocked": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "qty_disposed": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "disposition_reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation/FK error"
          }
        }
      }
    },
...
  }
}